+++ app/react-native-app/src/navigation/CaseStack.jsx
@@ -1,12 +1,10 @@
 import { Stack } from "./navigationStack";
 import { ScreenHeader } from "components/ScreenHeader";
-import { font, colors } from "style/style";
-import { CaseDetail } from "screens/CaseDetail";
-import { CaseList } from "screens/CaseList";
+import { colors } from "style";
 import { headerOptions } from "./HomeStack";
-import { FontAwesome, FontAwesome5, Ionicons } from "@expo/vector-icons";
+import { Ionicons } from "@expo/vector-icons";
 import { InitialConsultationStack } from "./InitialConsultationStack";
-import { CaseOverview } from "screens/CaseOverview";
+import { CaseList, CaseOverview, CaseDetail } from "screens/Case";
 import { SurveyForm } from "screens/Survey/SurveyForm";
 
 export const CaseStack = ({ navigation }) => {
@@ -16,15 +14,13 @@ export const CaseStack = ({ navigation }) => {
         name="CaseList"
         component={CaseList}
         options={{
-          headerBackground: () => (
-            <ScreenHeader title="Cases" fontSize={font.size["xl"]} />
-          ),
+          headerBackground: () => <ScreenHeader title="Cases" />,
           ...headerOptions,
           headerLeft: () => (
             <Ionicons
               name="menu-outline"
               size={24}
-              color={colors.light}
+              color={colors.palette.primary100}
               onPress={() => navigation.openDrawer()}
             />
           ),
@@ -37,9 +33,6 @@ export const CaseStack = ({ navigation }) => {
           headerBackground: () => (
             <ScreenHeader
               title={`Case Overview - ${route.params.caseData?.horse?.name}`}
-              fontSize={font.size.md}
-              fontWeight={400}
-              icon={<FontAwesome5 name="horse-head" size={18} />}
             />
           ),
           ...headerOptions,
@@ -52,9 +45,6 @@ export const CaseStack = ({ navigation }) => {
           headerBackground: () => (
             <ScreenHeader
               title={`Case detail - ${route.params.caseData?.horse?.name}`}
-              fontSize={font.size.md}
-              fontWeight={400}
-              icon={<FontAwesome5 name="horse-head" size={18} />}
             />
           ),
           ...headerOptions,
@@ -71,9 +61,6 @@ export const CaseStack = ({ navigation }) => {
           headerBackground: () => (
             <ScreenHeader
               title={`${route.params.surveyType?.name} - Survey form`}
-              fontSize={font.size.md}
-              fontWeight={400}
-              icon={<FontAwesome name="wpforms" size={18} />}
             />
           ),
           ...headerOptions,
