+++ app/react-native-app/src/screens/Case/CaseDetail.jsx
