+++ app/react-native-app/src/screens/InitialConsultation/index.jsx
@@ -1,0 +1,9 @@
+export * from "./InitialConsultationStepOne";
+export * from "./InitialConsultationStepTwo";
+export * from "./InitialConsultationStepThree";
+export * from "./InitialConsultationStepFour";
+export * from "./InitialConsultationStepFive";
+export * from "./InitialConsultationStepSix";
+export * from "./InitialConsultationStepSeven";
+export * from "./InitialConsultationStepEight";
+export * from "./Confirmation";
